+++ examples/data-sources/meshstack_platform_type/data-source.tf
@@ -1,0 +1,5 @@
+data "meshstack_platform_type" "example" {
+  metadata = {
+    name = "OPENSHIFT-4"
+  }
+}