+++ internal/provider/provider.go
@@ -154,6 +154,7 @@ func (p *MeshStackProvider) Resources(ctx context.Context) []func() resource.Res
 		NewPlatformResource,
 		NewPaymentMethodResource,
 		NewLocationResource,
+		NewPlatformTypeResource,
 	}
 }
 
@@ -174,6 +175,8 @@ func (p *MeshStackProvider) DataSources(ctx context.Context) []func() datasource
 		NewPlatformDataSource,
 		NewPaymentMethodDataSource,
 		NewIntegrationsDataSource,
+		NewPlatformTypesDataSource,
+		NewPlatformTypeDataSource,
 	}
 }